Add overall progress to backup script
`rsync -P` gives progress of the current file being transferred. Version 3.1 of
rsync added a new `--info` argument to show more detailed information on
transfers, which includes an option to show overall progress. For
transfers of lots of files (which this script is basically for) that is
preferable to the individual file progress.

Also shows a more detailed summary at the end.