Replace "bootstrap" with "update"
`update` performs the same actions as `bootstrap`, except that it also
stores the SHA of the file when it is installed.  This way, if the
current file and the stored SHA agree, then the file can be safely
replaced. If they do not agree, the user has modified the file, so it
will not be replaced without the 'install' flag.

As part of this, installing perl/python modules is no longer performed.
That is a job for my macos_setup repo, and not portable between OSs.

Further, update the only script that used Perl; replace the Perl code
with `date` commands in a bash script.